+++ back/fastats/src/main/java/org/sixbacks/fastats/statistics/repository/SectorRepository.java
@@ -3,6 +3,6 @@ package org.sixbacks.fastats.statistics.repository;
 import org.sixbacks.fastats.statistics.entity.Sector;
 import org.springframework.data.repository.ListCrudRepository;
 
-public interface SectorRepository extends ListCrudRepository<Sector, Integer> {
+public interface SectorRepository extends ListCrudRepository<Sector, Long> {
 
 }
